requirements.txt
playwright>=1.40.0
httpx>=0.25.0
openpyxl>=3.1.0
pyyaml>=6.0
nicegui>=2.0.0
fastapi>=0.110.0
uvicorn>=0.27.0
sse-starlette>=1.8.0
